requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pypdf
unstructured
docx2txt
openpyxl
pandas
networkx
langchain-chroma
msoffcrypto-tool